playground/requirements.txt
streamlit
graphviz
pandas
snowflake-connector-python
snowflake-snowpark-python